+++ KnowYourAudience-Quiz/app.py
@@ -1,20 +1,44 @@
 import os
 from dotenv import load_dotenv
 import streamlit as st
-from google import genai
-from google.genai import types
 from pathlib import Path
+import PyPDF2
+from google import genai
 
+# Gemini API Key
 load_dotenv()
-
-# Gemini API key
 GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
 
 # Initialize Gemini
 client = genai.Client(api_key=GEMINI_API_KEY)
 model = "gemini-2.5-flash-preview-05-20"
 
-# Define course files
+
+# Text extraction from PDF
+def extract_pdf_text(pdf_file) -> str:
+    text = ""
+    with open(pdf_file, "rb") as f:
+        pdf = PyPDF2.PdfReader(f)
+        for page in pdf.pages:
+            text += page.extract_text()
+    return text
+
+
+# Streamlit UI
+st.title("Welcome to your personalized study mentor!")
+
+# Display instructions
+st.info(
+    """
+     **Instructions:**  
+    1. Select a course from the list.  
+    2. Generate a practice question based on its content.  
+    3. Provide your answer in the text box.  
+    4. Get automated evaluation and feedback from Gemini!  
+    """
+)
+
+# List of courses with PDF files
 courses = {
     "Current Electricity": "courses/current-electricity-ncert-1-3.pdf",
     "Ray Optics": "courses/ray-optics-ncert.pdf",
@@ -22,92 +46,54 @@ courses = {
     "Matrices and Determinants": "courses/matrices-ncert.pdf",
 }
 
-# Streamlit UI
-st.title('Take your Learning Method to Next Level!')
 selected_course = st.selectbox(
-    "What course would you like to be assessed for?",
-    list(courses.keys()),
-    index=None,
-    placeholder="Select a course...",
-)
-st.write("You selected course:", selected_course)
+        "Select a Course to study!", 
+        list(courses.keys()), 
+        index=None
+    )
 
-def get_course_content(selected_course, courses):
-    # Retrieve and Encode the PDF byte
-    if not selected_course:
-        st.warning("Please select a course!")
 
-    course_path = Path(courses[selected_course])
-
-    return types.Part.from_bytes(
-            data= course_path.read_bytes(),
-            mime_type='application/pdf',
-    ) 
-
-course_content = None
 if selected_course:
-    course_content = get_course_content(selected_course, courses)
-
-def generate_question(course_content):
-    # Generate question using Gemini
-    prompt = f"Create a short question (fact-based, memory-based or reasoning based) by reading the document provided:\n\n{course_content}"
-    try:
-        response = client.models.generate_content(
-            model = model,
-            contents= [course_content, prompt],
-        )
-        question = response.text
-        return question
-    
-    except Exception as e:
-        st.error(f"Error generating question: {e}")
-        return None
-
-def evaluate_answer(question, user_answer):
-    if not question:
-        st.warning("No question available for evaluation.")
-        return None
-    
-    if not user_answer:
-        st.warning("Please enter your answer before evaluation.")
-        return None
-
-    eval_prompt = ("Evaluate the answer based on the course content. The evaluation componets should include whether the answer"
-    "was correct or not with a short reason, type of question(fact-based, memory-based or reasoning-based) and topic of question"
-    "from the course. Be short and consise.\n\n"f"Question:{question}\nUser Answer:{user_answer}\n")
-    try:
-        eval_response = client.models.generate_content(
-            model= model,
-            contents= [eval_prompt]
-        )
-        feedback = eval_response.text
-        st.write("### Evaluation and analysis:")
-        st.write(feedback)
-        return feedback
-    
-    except Exception as e:
-        st.error(f"Error evaluating answer:{e}")
-        return None
+    st.success(f"✅ {selected_course} selected.")
+    course_file = courses[selected_course]
+    course_content = extract_pdf_text(course_file)
 
 
-question_button = st.button("Generate Question")
-if 'question' not in st.session_state:
-    st.session_state.question = None
+    # Generate Question
+    if st.button("Generate Question") or st.session_state.get("generated", False):
+        st.session_state["generated"] = True
 
-if question_button or st.session_state.question is None:
-    if st.session_state.question is None:
-        st.session_state.question = generate_question(course_content)
+        if "generated_question" not in st.session_state:
+            # Generate question with Gemini
+            prompt = f"""Create a short question (fact-based, memory-based or reasoning based) by reading the document provided:"\n\n{course_content}"""    
 
-    if st.session_state.question:
+            response = client.models.generate_content(
+                model = model,
+                contents = [prompt],
+            )
+            st.session_state.generated_question = response.text.strip()
+
         st.write("### Question:")
-        st.write(st.session_state.question)
-        user_answer = st.text_input("Answer here:", key="user_answer")
-        eval_button = st.button("Evaluate Answer")
+        st.write(st.session_state.generated_question)
 
-        if eval_button:
-            # st.session_state.evaluate = True
-            st.write("Following analysis")
-    
-        # if eval_button and user_answer:
-            evaluate_answer(st.session_state.question, user_answer)
-            # st.session_state.evaluate = False 
+
+        # User Answer
+        user_answer = st.text_input("Enter your answer here:")
+
+        if st.button("Evaluate Answer") or st.session_state.get("evaluation_done", False):
+            st.session_state["evaluation_done"] = True
+
+            if user_answer:
+                eval_prompt = f"""Evaluate the answer based on the course content. Provide whether it is correct or not with a short reason, the type of question (fact-based, memory-based or reasoning-based) and the topic of the question. Be short and concise.
+                    Question: {st.session_state.generated_question}
+                    User Answer: {user_answer}
+                    Course Content: {course_content}
+                    """
+                response = client.models.generate_content(
+                    model = model,
+                    contents = [eval_prompt],
+                )
+                st.write("### Evaluation and Analysis:")
+                st.write(response.text)
+            else:
+                st.error("⚡ Please provide your answer first.")
